backend/Bariki/requirements.txt
django
djangorestframework
django-cors-headers
django_daraja